refactor: move guest data to own file
